* #9979: Don't do runtime assertions on data.

+++ daemon/src/audio/audiortp/audio_rtp_session.cpp
@@ -156,7 +156,7 @@ void AudioRtpSession::receiveSpeakerData()
         return;
 
     unsigned char* spkrDataIn = (unsigned char*) adu->getData(); // data in char
-    unsigned int size = adu->getSize(); // size in char
+    size_t size = adu->getSize(); // size in char
 
     // DTMF over RTP, size must be over 4 in order to process it as voice data
     if (size > 4)
